src/bootstrap: add UEFI target support
The Rust compiler assumes that a certain set of built-in functions
is available for all systems. These functions include
memcpy/memset/memmove etc. They are provide by C library. But UEFI
doesn't have C library. So add `compiler-builtins-mem` feature to
compiler-builtins to enable these functions.

I'm not sure if it is a good approach. Any better idea is welcome.